Ghost Train!
A rudimentary form of a ghost train is in the works. When complete, it will auto-build and auto-destroy a minecart track to the nearest biome of your choosing.

Willo Wisp is a guiding mob which generates light using Bright Space blocks (a reliable alternative to dynamic lighting.) This mob will eventually be responsible for handing the generation and destruction of the minecart tracks.