Ghost Train is now mostly functional.
It still wants a GUI, but is presently user-friendly enough to be used without a console. When the light turns green, right click to fast-travel 'somewhere else'.

Presently finds paths in order of biomeID, will eventually find paths in order of closest biomes first.

Am experimenting with weighted pathfinding. Presently putting a tiny weight (1%) on closer nodes. Seems to speed up pathfinding over massive distances DRASTICALLY.

Augmented MinecartPathThread to communicate with IMinecartPathedEntity when it needs chunks loaded. This is pretty useful and should probably be put into the regular pathfinding algorithm.

Also, have been finding A* a bit faster than JPS? Not sure what's up with that, but will leave it as it is for now.